requirements.txt
fastapi
uvicorn
google-genai
httpx
python-dotenv
pydantic
pydantic-settings
